add a text box for the sampling frequency setting

+++ fusion/sensor_fusion.py
@@ -48,10 +48,10 @@ class Sensors:
         self.SAMPLING_TIME = 1 / self.SAMPLING_FREQUENCY_HZ
         self.SAVE_DATA_DIR = config.save_data_dir
         self.SAVE_BUF_CSVDATA_PATH = self.SAVE_DATA_DIR + "/" + "measurement_raw_data.csv"
-        self.SEQUENCE_LENGTH = config.sequence_length # Windows size [s]
+        self.SEQUENCE_LENGTH = int(config.sequence_length) # Windows size [s]
         # Buffer size is determined by the relation of sequence length and sampling frequency
         # Buffer secures data for SEQUENCE_LENGTH[s]
-        self.MAX_DATA_BUF_LEN = self.SEQUENCE_LENGTH * self.SAMPLING_FREQUENCY_HZ
+        self.MAX_DATA_BUF_LEN = int(self.SEQUENCE_LENGTH * self.SAMPLING_FREQUENCY_HZ)
         self.FPASS = config.filter_params.fpass
         self.FSTOP = config.filter_params.fstop
         self.GPASS = config.filter_params.gpass
